hf_upload/src/seve/link.py
"""
SEVE-Link: Integração ERP e Sincronização

Componente responsável por:
- Conectividade com ERPs (SAP, Oracle, TOTVS, Microsoft)
- API Gateway unificado
- Sincronização em tempo real
- Gerenciamento de webhooks
- Integração de sistemas externos
"""

from typing import Dict, List, Any, Optional
from dataclasses import dataclass
from enum import Enum
import asyncio
import aiohttp


class ERPType(Enum):
    """Tipos de ERP suportados"""
    SAP = "sap"
    ORACLE = "oracle"
    TOTVS = "totvs"
    MICROSOFT_DYNAMICS = "microsoft_dynamics"
    LINX = "linx"
    GENERIC = "generic"


class SyncStatus(Enum):
    """Status de sincronização"""
    SUCCESS = "success"
    PENDING = "pending"
    FAILED = "failed"
    PARTIAL = "partial"


@dataclass
class ERPConnection:
    """Configuração de conexão ERP"""
    erp_type: ERPType
    endpoint: str
    credentials: Dict[str, str]
    timeout: int = 30
    retry_attempts: int = 3


@dataclass
class SyncResult:
    """Resultado de sincronização"""
    status: SyncStatus
    records_synced: int
    errors: List[str]
    timestamp: float
    duration: float


class SEVELink:
    """
    Sistema de integração ERP e sincronização
    
    Fornece conectividade unificada com diferentes ERPs
    e sistemas externos para sincronização de dados.
    """
    
    def __init__(self, config: Optional[Dict[str, Any]] = None):
        """
        Inicializa o SEVE-Link
        
        Args:
            config: Configurações de integração
        """
        self.config = config or {}
        self.erp_connector = ERPConnector()
        self.api_gateway = APIGateway()
        self.data_sync = DataSynchronizationEngine()
        self.webhook_manager = WebhookManager()
        self.connections: Dict[str, ERPConnection] = {}
    
    def add_erp_connection(
        self,
        connection_id: str,
        erp_connection: ERPConnection
    ) -> bool:
        """
        Adiciona conexão com ERP
        
        Args:
            connection_id: ID da conexão
            erp_connection: Configuração da conexão
            
        Returns:
            True se conexão foi adicionada com sucesso
        """
        try:
            self.connections[connection_id] = erp_connection
            return True
        except Exception:
            return False
    
    async def sync_transaction(
        self,
        transaction_data: Dict[str, Any],
        connection_id: str
    ) -> SyncResult:
        """
        Sincroniza transação com ERP
        
        Args:
            transaction_data: Dados da transação
            connection_id: ID da conexão ERP
            
        Returns:
            Resultado da sincronização
        """
        if connection_id not in self.connections:
            return SyncResult(
                status=SyncStatus.FAILED,
                records_synced=0,
                errors=["Connection not found"],
                timestamp=0.0,
                duration=0.0
            )
        
        connection = self.connections[connection_id]
        
        # Enviar transação para ERP
        erp_response = await self.erp_connector.send_transaction(
            transaction_data, connection
        )
        
        # Atualizar inventário
        inventory_update = await self.data_sync.update_inventory(
            transaction_data
        )
        
        # Notificar via webhook
        await self.webhook_manager.notify_stakeholders(
            transaction_data, erp_response
        )
        
        return SyncResult(
            status=SyncStatus.SUCCESS,
            records_synced=1,
            errors=[],
            timestamp=0.0,
            duration=0.1
        )
    
    async def sync_inventory(self, connection_id: str) -> SyncResult:
        """
        Sincroniza inventário com ERP
        
        Args:
            connection_id: ID da conexão ERP
            
        Returns:
            Resultado da sincronização
        """
        if connection_id not in self.connections:
            return SyncResult(
                status=SyncStatus.FAILED,
                records_synced=0,
                errors=["Connection not found"],
                timestamp=0.0,
                duration=0.0
            )
        
        connection = self.connections[connection_id]
        
        # Sincronizar inventário
        inventory_data = await self.erp_connector.get_inventory(connection)
        sync_result = await self.data_sync.sync_inventory_data(inventory_data)
        
        return sync_result
    
    def get_connection_status(self, connection_id: str) -> Dict[str, Any]:
        """Retorna status da conexão"""
        if connection_id not in self.connections:
            return {"status": "not_found"}
        
        connection = self.connections[connection_id]
        return {
            "status": "active",
            "erp_type": connection.erp_type.value,
            "endpoint": connection.endpoint,
            "last_sync": "2025-01-27T10:00:00Z"
        }
    
    def get_all_connections(self) -> Dict[str, Dict[str, Any]]:
        """Retorna todas as conexões"""
        return {
            conn_id: {
                "erp_type": conn.erp_type.value,
                "endpoint": conn.endpoint,
                "status": "active"
            }
            for conn_id, conn in self.connections.items()
        }


class ERPConnector:
    """Conector para ERPs"""
    
    async def send_transaction(
        self,
        transaction_data: Dict[str, Any],
        connection: ERPConnection
    ) -> Dict[str, Any]:
        """Envia transação para ERP"""
        # Placeholder para envio real
        return {
            "status": "success",
            "transaction_id": "ERP_12345",
            "timestamp": "2025-01-27T10:00:00Z"
        }
    
    async def get_inventory(self, connection: ERPConnection) -> List[Dict[str, Any]]:
        """Recupera inventário do ERP"""
        # Placeholder para recuperação real
        return [
            {
                "product_id": "PROD_001",
                "quantity": 100,
                "price": 10.50,
                "last_updated": "2025-01-27T10:00:00Z"
            }
        ]


class APIGateway:
    """Gateway de API unificado"""
    
    def __init__(self):
        self.routes = {}
        self.middleware = []
    
    def add_route(self, path: str, handler: Any) -> None:
        """Adiciona rota ao gateway"""
        self.routes[path] = handler
    
    async def handle_request(self, path: str, data: Dict[str, Any]) -> Dict[str, Any]:
        """Processa requisição através do gateway"""
        if path in self.routes:
            return await self.routes[path](data)
        return {"error": "Route not found"}


class DataSynchronizationEngine:
    """Motor de sincronização de dados"""
    
    async def update_inventory(self, transaction_data: Dict[str, Any]) -> Dict[str, Any]:
        """Atualiza inventário local"""
        # Placeholder para atualização real
        return {
            "status": "updated",
            "products_updated": len(transaction_data.get("products", [])),
            "timestamp": "2025-01-27T10:00:00Z"
        }
    
    async def sync_inventory_data(self, inventory_data: List[Dict[str, Any]]) -> SyncResult:
        """Sincroniza dados de inventário"""
        return SyncResult(
            status=SyncStatus.SUCCESS,
            records_synced=len(inventory_data),
            errors=[],
            timestamp=0.0,
            duration=0.05
        )


class WebhookManager:
    """Gerenciador de webhooks"""
    
    def __init__(self):
        self.webhooks = []
    
    def add_webhook(self, url: str, events: List[str]) -> None:
        """Adiciona webhook"""
        self.webhooks.append({
            "url": url,
            "events": events,
            "active": True
        })
    
    async def notify_stakeholders(
        self,
        transaction_data: Dict[str, Any],
        erp_response: Dict[str, Any]
    ) -> None:
        """Notifica stakeholders via webhook"""
        # Placeholder para notificação real
        for webhook in self.webhooks:
            if webhook["active"]:
                # Enviar notificação
                pass
